refactor: separate responsibilities by feature in controller as suggested in code review
