+++ controllers/create_controller.go
@@ -1,0 +1,28 @@
+package controllers
+
+import (
+	"net/http"
+
+	"github.com/dornascarol/api-go-gin/database"
+	"github.com/dornascarol/api-go-gin/models"
+	"github.com/gin-gonic/gin"
+)
+
+func CreateNewSinger(c *gin.Context) {
+	var singer models.Singer
+
+	if err := c.ShouldBindJSON(&singer); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error()})
+		return
+	}
+
+	if err := models.ValidateSingerData(&singer); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error()})
+		return
+	}
+
+	database.DB.Create(&singer)
+	c.JSON(http.StatusOK, singer)
+}
